requirements.txt
# Autonomous Business Logic Orchestrator - Dependencies
# Python 3.10+

# CrewAI & orchestration
crewai>=0.80.0
crewai-tools>=0.14.0

# LangChain & LLM
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Vector store (Pinecone for RAG)
pinecone-client>=5.0.0

# Data validation & config
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database (ERP simulation)
# SQLite is in stdlib

# HTTP & async
httpx>=0.27.0
aiohttp>=3.9.0

# Logging & dev
structlog>=24.0.0
rich>=13.0.0
